1、同步ruoyi最新代码

+++ ruoyi-common/src/main/java/com/ruoyi/common/core/page/TableDataInfo.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 /**
  * 表格分页数据对象
- * 
+ *
  * @author ruoyi
  */
 public class TableDataInfo implements Serializable
@@ -33,11 +33,11 @@ public class TableDataInfo implements Serializable
 
     /**
      * 分页
-     * 
+     *
      * @param list 列表数据
      * @param total 总记录数
      */
-    public TableDataInfo(List<?> list, int total)
+    public TableDataInfo(List<?> list, long total)
     {
         this.rows = list;
         this.total = total;
